requirements.txt
grpcio~=1.29.0
opentelemetry-api>=0.9b0
opentelemetry-sdk>=0.9b0
opentelemetry-ext-otlp>=0.9b0
opentelemetry-instrumentation>=0.9b0